extract delimsizing functions into their own file
